+++ examples/aptos-friend/frontend/main.tsx
@@ -7,10 +7,12 @@ import App from "@/App.tsx";
 // Internal components
 import { Toaster } from "@/components/ui/toaster.tsx";
 import { WalletProvider } from "@/components/WalletProvider.tsx";
+import { WrongNetworkAlert } from "@/components/WrongNetworkAlert";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <WalletProvider>
+      <WrongNetworkAlert />
       <App />
       <Toaster />
     </WalletProvider>